clean up the vstd_build module of cargo-verus
